RSC convert admin interactive pages: sections, beds, holidays
