+++ src/app/(admin)/admin/sections/page.tsx
@@ -1,181 +1,12 @@
-"use client";
+import { fetchSectionVisibility } from "@/lib/data/branding";
+import { mergeSectionVisibility } from "@/lib/section-visibility";
+import { requireTenant } from "@/lib/tenant";
+import SectionsForm from "./_sections-form";
 
-import { AlertCircle, Loader2, Save, ToggleLeft, ToggleRight } from "lucide-react";
-import { useState } from "react";
-import { Badge } from "@/components/ui/badge";
-import { Breadcrumb } from "@/components/ui/breadcrumb";
-import { Button } from "@/components/ui/button";
-import { Card, CardContent, CardHeader, CardTitle, CardDescription } from "@/components/ui/card";
-import { useAsyncData } from "@/lib/hooks/use-async-data";
-import {
-  sectionDefinitions,
-  defaultSectionVisibility,
-  mergeSectionVisibility,
-  type SectionVisibility,
-  type SectionKey,
-} from "@/lib/section-visibility";
+export default async function SectionsPage() {
+  const tenant = await requireTenant();
+  const saved = await fetchSectionVisibility(tenant.clinicId);
+  const initialVisibility = mergeSectionVisibility(saved);
 
-export default function SectionsPage() {
-  const {
-    data: initialData,
-    loading,
-    error,
-  } = useAsyncData(
-    (signal) =>
-      fetch("/api/branding", { signal })
-        .then((r) => {
-          if (!r.ok) throw new Error(`Failed to load sections (${r.status})`);
-          return r.json();
-        })
-        .then((json) => {
-          const data = json.data ?? json;
-          return mergeSectionVisibility(data.section_visibility);
-        }),
-    defaultSectionVisibility,
-  );
-  const [visibility, setVisibility] = useState<SectionVisibility>(defaultSectionVisibility);
-  const [savedState, setSavedState] = useState<SectionVisibility>(defaultSectionVisibility);
-  const [initialized, setInitialized] = useState(false);
-  const [saving, setSaving] = useState(false);
-
-  // Sync initial data once loaded
-  if (!initialized && !loading && !error) {
-    setVisibility(initialData);
-    setSavedState(initialData);
-    setInitialized(true);
-  }
-
-  const toggle = (key: SectionKey) => {
-    const def = sectionDefinitions.find((s) => s.key === key);
-    if (def?.alwaysOn) return;
-    setVisibility((prev) => ({ ...prev, [key]: !prev[key] }));
-  };
-
-  const handleSave = async () => {
-    setSaving(true);
-    try {
-      const res = await fetch("/api/branding", {
-        method: "PUT",
-        headers: { "Content-Type": "application/json" },
-        body: JSON.stringify({ section_visibility: visibility }),
-      });
-      if (!res.ok) {
-        const err = await res.json().catch(() => null);
-        alert(err?.error ?? "Failed to save section settings");
-        return;
-      }
-      setSavedState(visibility);
-    } catch {
-      alert("Failed to save section settings");
-    } finally {
-      setSaving(false);
-    }
-  };
-
-  const hasChanges = JSON.stringify(visibility) !== JSON.stringify(savedState);
-
-  const enabledCount = Object.values(visibility).filter(Boolean).length;
-
-  if (loading) {
-    return (
-      <div>
-        <h1 className="text-2xl font-bold mb-6">Section Control</h1>
-        <p className="text-muted-foreground flex items-center gap-2">
-          <Loader2 className="h-4 w-4 animate-spin" />
-          Loading section settings...
-        </p>
-      </div>
-    );
-  }
-
-  if (error) {
-    return (
-      <div>
-        <h1 className="text-2xl font-bold mb-6">Section Control</h1>
-        <div className="rounded-lg border border-destructive/50 bg-destructive/10 p-4 flex items-start gap-3">
-          <AlertCircle className="h-5 w-5 text-destructive mt-0.5" />
-          <div>
-            <p className="font-medium text-destructive">Failed to load section settings</p>
-            <p className="text-sm text-muted-foreground mt-1">{error.message}</p>
-          </div>
-        </div>
-      </div>
-    );
-  }
-
-  return (
-    <div>
-      <Breadcrumb items={[{ label: "Admin", href: "/admin/dashboard" }, { label: "Sections" }]} />
-      <div className="flex items-center justify-between mb-6">
-        <div>
-          <h1 className="text-2xl font-bold">Section Control</h1>
-          <p className="text-sm text-muted-foreground">
-            Toggle sections on or off on your public homepage. The order stays fixed per template.
-          </p>
-          <p className="text-xs text-muted-foreground mt-1">
-            {enabledCount} of {sectionDefinitions.length} sections enabled
-          </p>
-        </div>
-        <Button onClick={handleSave} disabled={saving || !hasChanges}>
-          <Save className="h-4 w-4 mr-2" />
-          {saving ? "Saving..." : hasChanges ? "Save Changes" : "Saved"}
-        </Button>
-      </div>
-
-      <Card>
-        <CardHeader>
-          <CardTitle className="text-base">Homepage Sections</CardTitle>
-          <CardDescription>
-            Sections are displayed in the order listed below. Toggle each section to show or hide it
-            on your public site.
-          </CardDescription>
-        </CardHeader>
-        <CardContent>
-          <div className="divide-y">
-            {sectionDefinitions.map((section) => {
-              const isOn = visibility[section.key];
-              const isAlwaysOn = section.alwaysOn === true;
-
-              return (
-                <div
-                  key={section.key}
-                  className={`flex items-center justify-between py-4 first:pt-0 last:pb-0 ${
-                    !isOn && !isAlwaysOn ? "opacity-50" : ""
-                  }`}
-                >
-                  <div className="space-y-0.5">
-                    <div className="flex items-center gap-2">
-                      <span className="text-sm font-medium">{section.label}</span>
-                      {isAlwaysOn && (
-                        <Badge variant="secondary" className="text-[10px]">
-                          Always on
-                        </Badge>
-                      )}
-                    </div>
-                    <p className="text-xs text-muted-foreground">{section.description}</p>
-                  </div>
-
-                  <button
-                    type="button"
-                    onClick={() => toggle(section.key)}
-                    disabled={isAlwaysOn}
-                    className={`flex-shrink-0 transition-colors ${
-                      isAlwaysOn ? "cursor-not-allowed opacity-60" : "cursor-pointer"
-                    }`}
-                    aria-label={`Toggle ${section.label}`}
-                  >
-                    {isOn ? (
-                      <ToggleRight className="h-8 w-8 text-primary" />
-                    ) : (
-                      <ToggleLeft className="h-8 w-8 text-muted-foreground" />
-                    )}
-                  </button>
-                </div>
-              );
-            })}
-          </div>
-        </CardContent>
-      </Card>
-    </div>
-  );
+  return <SectionsForm initialVisibility={initialVisibility} />;
 }
